requirements.txt
numpy==1.24.2
opencv-python-headless==4.7.0.72
Pillow==9.5.0
#Pillow==10.0.1
XlsxWriter==3.1.4
streamlit==1.27.1